test/requirement.txt
flask
flasgger
tensorflow
pillow
numpy
keras
